+++ src/Painter.h
@@ -86,7 +86,7 @@ public:
 	
 	/** Draw a string at position x, y (y is the baseline)
 	    This is just for fast drawing */
-	PainterBase & text(int x, int y, char const * str, int l,
+	PainterBase & text(int x, int y, char const * str, size_t l,
 			   LyXFont const & f);
 	
 	/// Draw a char at position x, y (y is the baseline)
